# website/client/nswag/openapi.yaml
openapi: 3.0.0
info:
  title: Swagger Lumileds
  description: Lumileds Quiz API
  contact:
    email: thinhtran@techkids.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.3.0
servers:
- url: https://lumileds.techkids.io/api
...
paths:
  /auth/register:
    post:
      tags:
      - auth
      summary: Register new account
      operationId: register
      requestBody:
        $ref: '#/components/requestBodies/RegisterDTO'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        400:
          description: invalid input
        500:
          description: internal server error
          content:
            application/json:
              schema:
                type: string
  /auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login
      requestBody:
        $ref: '#/components/requestBodies/LoginDTO'
      responses:
        200:
          description: successful operation
          content:
            application/json; charset=utf-8:
              schema:
                type: string
        400:
          description: invalid input
        500:
          description: internal server error
          content:
            application/json:
              schema:
                type: string
  /auth/forgotpassword:
    post:
      tags:
      - auth
      summary: Forgot password
      operationId: forgotPassword
      requestBody:
        $ref: '#/components/requestBodies/ForgotPasswordDTO'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        500:
          description: internal server error
          content:
            application/json:
              schema:
                type: string
      security:
      - Bearer: []
  /auth/resetpassword/{token}:
    post:
      tags:
      - auth
      summary: Reset password
      operationId: resetPassword
      parameters:
      - name: token
        in: path
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/ResetPasswordDTO'
      responses:
        200:
          description: Success! Your password has been changed.
          content:
            application/json:
              schema:
                type: string
        400:
          description: invalid input
        500:
          description: internal server error
          content:
            application/json:
              schema:
                type: string
...
  /quizzes:
    get:
      tags:
      - quizzes
      summary: Find quizzes
      operationId: getAllQuizzes
      parameters:
      - name: token
        in: header
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: searchTerm
        in: query
        description: search by name
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: state
        in: query
        description: search by quiz state
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: pageIndex
        in: query
        description: the current page index
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: pageSize
        in: query
        description: items in each page
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: sortBy
        in: query
        description: field name to sort
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: sortAsc
        in: query
        description: sort ascending or descending
        required: false
        style: form
        explode: true
        schema:
          type: boolean
      responses:
        200:
          description: successful operation
          content:
            application/json; charset=utf-8:
              schema:
                type: string
        400:
          description: invalid input
        401:
          description: Token is invalid or expired
        500:
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizzesDTO'
      security:
      - Bearer: []
    post:
      tags:
      - quizzes
      summary: Create new quiz
      operationId: createQuiz
      parameters:
      - name: token
        in: header
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/CreateQuizDTO'
      responses:
        200:
          description: successful operation
          content:
            application/json; charset=utf-8:
              schema:
                type: string
        400:
          description: invalid input
        401:
          description: Token is invalid or expired
        500:
          description: internal server error
          content:
            application/json:
              schema:
                type: string
      security:
      - Bearer: []
  /quizzes/{quizId}:
    get:
      tags:
      - quizzes
      summary: Get quiz by id
      operationId: findQuizz
      parameters:
      - name: token
        in: header
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: quizId
        in: path
        description: quizId
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        401:
          description: Token is invalid or expired
        500:
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizDTO'
      security:
      - Bearer: []
    put:
      tags:
      - quizzes
      summary: Update quiz
      operationId: updateQuiz
      parameters:
      - name: token
        in: header
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: quizId
        in: path
        description: quizId
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/CreateQuizDTO'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        400:
          description: invalid input
        401:
          description: Token is invalid or expired
        500:
          description: internal server error
          content:
            application/json:
              schema:
                type: string
      security:
      - Bearer: []
    delete:
      tags:
      - quizzes
      summary: Remove quiz
      operationId: deleteQuiz
      parameters:
      - name: token
        in: header
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: quizId
        in: path
        description: quizId
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        401:
          description: Token is invalid or expired
        500:
          description: internal server error
          content:
            application/json:
              schema:
                type: string
      security:
      - Bearer: []
  /quizzes/{quizId}/activate:
    get:
      tags:
      - quizzes
      summary: Activate quiz
      operationId: activateQuiz
      parameters:
      - name: token
        in: header
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: quizId
        in: path
        description: quizId
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        401:
          description: Token is invalid or expired
        500:
          description: internal server error
          content:
            application/json:
              schema:
                type: string
      security:
      - Bearer: []
  /quizzes/{quizId}/scoreboard:
    get:
      tags:
      - quizzes
      summary: Get quiz's scoreboard
      operationId: getQuizScoreboard
      parameters:
      - name: token
        in: header
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: quizId
        in: path
        description: quizId
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: userId
        in: query
        description: userId
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: pageIndex
        in: query
        description: the current page index
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: pageSize
        in: query
        description: items in each page
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        401:
          description: Token is invalid or expired
        500:
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreboardDTO'
      security:
      - Bearer: []
  /quizzes/submitanswers:
    post:
      tags:
      - quizzes
      summary: Submit answer
      operationId: submitQuizAnswers
      parameters:
      - name: token
        in: header
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/SubmitQuizAnswersDTO'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        400:
          description: invalid input
        401:
          description: Token is invalid or expired
        500:
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitQuizAnswersResultDTO'
      security:
      - Bearer: []
  /uploadImage:
    post:
      tags:
      - uploadImage
      summary: Upload image
      operationId: uploadImage
      parameters:
      - name: token
        in: header
        description: token
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/body'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        401:
          description: Token is invalid or expired
        500:
          description: internal server error
          content:
            application/json:
              schema:
                type: string
                description: image url
                example: /images/8034455b-1076-4cac-a732-c413c234b666.jpg
      security:
      - Bearer: []
...
components:
  schemas:
    RegisterDTO:
      required:
      - email
      - fullName
      - password
      - username
      type: object
      properties:
        fullName:
          type: string
          example: anhvu
        username:
          type: string
          example: thinhtran
        email:
          type: string
          example: thinhtran@techkids.io
        password:
          type: string
          example: password
    LoginDTO:
      required:
      - password
      - rememberMe
      - usernameOrEmail
      type: object
      properties:
        usernameOrEmail:
          type: string
          example: thinhtran
        password:
          type: string
          example: password
        rememberMe:
          type: boolean
...
    ForgotPasswordDTO:
      type: object
      properties:
        email:
          type: string
          example: thinhtran@techkids.io
    ResetPasswordDTO:
      type: object
      properties:
        password:
          type: string
          example: Abc@12345
        email:
          type: string
          example: Abc@gmail.com
...
    QuizzesDTO:
      type: object
      properties:
        total:
          type: integer
          format: int32
          example: 100
        data:
          type: array
          items:
            $ref: '#/components/schemas/QuizzesDTO_data'
    CreateQuizDTO:
      type: object
      properties:
        coverImageUrl:
          type: string
          example: /images/121215.jpg
        title:
          type: string
          example: the first quiz
        description:
          type: string
          example: quiz description
        state:
          type: string
          example: DRAFT/PUBLISHED
        beacon:
          $ref: '#/components/schemas/QuizzesDTO_beacon'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/CreateQuizDTO_questions'
    QuizDTO:
      type: object
      properties:
        questionCount:
          type: integer
          format: int32
          example: 10
        _id:
          type: string
          example: 8034455b-1076-4cac-a732-c413c234b666
        coverImageUrl:
          type: string
          example: /images/121215.jpg
        title:
          type: string
          example: the first quiz
        description:
          type: string
          example: quiz description
        state:
          type: string
          example: DRAFT/PUBLISHED
        createdAt:
          type: integer
          format: int32
        beacon:
          $ref: '#/components/schemas/QuizzesDTO_beacon'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizzesDTO_questions'
    ScoreboardDTO:
      type: object
      properties:
        total:
          type: integer
          format: int32
          example: 100
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScoreboardDTO_data'
    SubmitQuizAnswersDTO:
      type: object
      properties:
        userId:
          type: string
          example: 8034455b-1076-4cac-a732-c413c234b666
        quizId:
          type: string
          example: 8034455b-1076-4cac-a732-c413c234b666
        answers:
          type: array
          items:
            $ref: '#/components/schemas/SubmitQuizAnswersDTO_answers'
    SubmitQuizAnswersResultDTO:
      type: object
      properties:
        correctAnswerCount:
          type: integer
          format: int32
          example: 6
        points:
          type: integer
          format: int32
          example: 6
    body:
      type: object
      properties:
        fileName:
          type: string
          format: binary
...
    QuizzesDTO_beacon:
      type: object
      properties:
        isActive:
          type: boolean
        uuid:
          type: string
          example: 12411g24215125g12g12
        major:
          type: string
          example: "10"
        minor:
          type: string
          example: "10"
    QuizzesDTO_answers:
      type: object
      properties:
        id:
          type: string
          example: 0/1/2/3
        décription:
          type: string
          example: answer 1
        isCorrect:
          type: boolean
    QuizzesDTO_questions:
      type: object
      properties:
        id:
          type: string
          example: 8034455b-1076-4cac-a732-c413c234b666
        coverType:
          type: string
          example: IMAGE/VIDEO
        coverUrl:
          type: string
          example: images/8034455b-1076-4cac-a732-c413c234b666.jpg
        description:
          type: string
          example: question description
        answers:
          type: array
          items:
            $ref: '#/components/schemas/QuizzesDTO_answers'
    QuizzesDTO_data:
      type: object
      properties:
        questionCount:
          type: integer
          format: int32
          example: 10
        _id:
          type: string
          example: 8034455b-1076-4cac-a732-c413c234b666
        coverImageUrl:
          type: string
          example: /images/121215.jpg
        title:
          type: string
          example: the first quiz
        description:
          type: string
          example: quiz description
        state:
          type: string
          example: DRAFT/PUBLISHED
        createdAt:
          type: integer
          format: int32
        beacon:
          $ref: '#/components/schemas/QuizzesDTO_beacon'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizzesDTO_questions'
    CreateQuizDTO_answers:
      type: object
      properties:
        id:
          type: string
          example: 0/1/2/3
        description:
          type: string
          example: answer 1
        isCorrect:
          type: boolean
    CreateQuizDTO_questions:
      type: object
      properties:
        id:
          type: string
          example: 8034455b-1076-4cac-a732-c413c234b666
        coverType:
          type: string
          example: IMAGE/VIDEO
        coverUrl:
          type: string
          example: images/8034455b-1076-4cac-a732-c413c234b666.jpg
        description:
          type: string
          example: question description
        answers:
          type: array
          items:
            $ref: '#/components/schemas/CreateQuizDTO_answers'
    ScoreboardDTO_data:
      type: object
      properties:
        userId:
          type: string
          example: 8034455b-1076-4cac-a732-c413c234b666
        name:
          type: string
          example: Thinh Tran
        points:
          type: integer
          format: int32
          example: 1000
    SubmitQuizAnswersDTO_answers:
      type: object
      properties:
        questionId:
          type: string
          example: 8034455b-1076-4cac-a732-c413c234b666
        answerId:
          type: number
  requestBodies:
    RegisterDTO:
      description: Register params
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RegisterDTO'
      required: true
...
    LoginDTO:
      description: Login params
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LoginDTO'
      required: true
    ForgotPasswordDTO:
      description: Forgot Password params
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ForgotPasswordDTO'
      required: true
    ResetPasswordDTO:
      description: Reset Password params
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResetPasswordDTO'
      required: true
...
    CreateQuizDTO:
      description: Cancel Redeem points
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateQuizDTO'
      required: true
...
    SubmitQuizAnswersDTO:
      description: Cancel Redeem points
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SubmitQuizAnswersDTO'
      required: true
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header